requirements.txt
discord==2.3.2
google-generativeai==0.6.0
python-dotenv==1.0.1
requests==2.32.3
grpcio==1.71.0
